Cleanup and code documentation
