+++ src/presets/src/Presets.h
@@ -23,7 +23,7 @@
 #include <iostream>
 #include <string>
 
-#include "SimVars.h"
+#include "LightingSimVars.h"
 
 using namespace std;
 
@@ -34,33 +34,72 @@ class Presets {
   bool isConnected = false;
 
   SimulationData simulationData = {};
-  SimVars* simVars;
+  LightingSimVars* simVars;
 
-  // Initializes the connection to SimConnect
-  // <returns>True if successful, false otherwise.</returns>
+  /**
+   * Initialize and connect the SimConnect connection.
+   * @return true on success, false on failure
+   */
   bool initializeSimConnect();
 
  public:
 
-  // Initializes the PRESETS control
-  // <returns>True if successful, false otherwise.</returns>
-  bool initializePRESETS();
+  /**
+   * Initialize the gauge (instead of a constructor).
+   * Sets up data for the gauge and also connect to SimConnect.
+   * @return true if SimConnect was successfully connected, false otherwise.
+   */
+  bool initialize();
 
-  // Callback used to update the PRESETS at each tick (dt)
-  // <returns>True if successful, false otherwise.</returns>
+  /**
+   * Callback used to update the PRESETS at each tick (dt).
+   * This is used to execute every action and task required to update the gauge.
+   * @param deltaTime The time since the last tick
+   * @return True if successful, false otherwise.
+   */
   bool onUpdate(double deltaTime);
 
+  /**
+   * Request SimConnect data for the prepared data structure.
+   * (SimConnect_RequestDataOnSimObject)
+   * @return true if successful, false otherwise
+   */
   bool simConnectRequestData();
 
+  /**
+   * Retrieve SimConnect data for the prepared data structure.
+   * (SimConnect_GetNextDispatch)
+   * @return true if successful, false otherwise   */
   bool simConnectReadData();
 
+  /**
+   * Process received SimConnect messages.
+   * https://docs.flightsimulator.com/html/Programming_Tools/SimConnect/API_Reference/General/SimConnect_GetNextDispatch.htm?rhhlterm=SimConnect_GetNextDispatch&rhsearch=SimConnect_GetNextDispatch
+   * @param pData Pointer to a pointer to a data buffer, initially to be treated as a
+   * SIMCONNECT_RECV structure. If you are going to make a copy of the data buffer (which is
+   * maintained by the SimConnect client library) make sure that the defined buffer is large
+   * enough (the size of the returned data structure is one member of the SIMCONNECT_RECV structure.
+   * @param cbData Pointer to the size of the data buffer, in bytes.
+   */
   void simConnectProcessDispatchMessage(SIMCONNECT_RECV* pData, DWORD* cbData);
 
+  /**
+   * Procees received CimConnect sim object data
+   * https://docs.flightsimulator.com/html/Programming_Tools/SimConnect/API_Reference/Structures_And_Enumerations/SIMCONNECT_RECV_SIMOBJECT_DATA.htm?rhhlterm=SIMCONNECT_RECV_SIMOBJECT_DATA&rhsearch=SIMCONNECT_RECV_SIMOBJECT_DATA
+   * @param data pointer to the data object
+   */
   void simConnectProcessSimObjectData(const SIMCONNECT_RECV_SIMOBJECT_DATA* data);
 
-  // Kills the PRESETS and unregisters all LVars
-  // <returns>True if successful, false otherwise.</returns>
-  bool killPRESETS();
+  /**
+   * Kills the PRESETS and unregisters all LVars
+   * @return True if successful, false otherwise.
+   */
+  bool shutdown();
 
+  /**
+   * Returns human readable string representation of an exception.
+   * @param exception https://docs.flightsimulator.com/html/Programming_Tools/SimConnect/API_Reference/Structures_And_Enumerations/SIMCONNECT_EXCEPTION.htm?rhhlterm=SIMCONNECT_EXCEPTION&rhsearch=SIMCONNECT_EXCEPTION
+   * @return string representation of exception.
+   */
   std::string getSimConnectExceptionString(SIMCONNECT_EXCEPTION exception);
 };
